Added search movie

+++ src/main/java/com/socialbox/model/Tmdb.java
@@ -12,18 +12,18 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class Tmdb {
-    private boolean adult;
-    private String backdrop_path;
-    private List<Integer> genre_ids;
-    private String id;
-    private String original_language;
-    private String original_title;
-    private String overview;
-    private String popularity;
-    private String poster_path;
-    private String release_date;
-    private String title;
-    private boolean video;
-    private int vote_average;
-    private int vote_count;
+  private boolean adult;
+  private String backdrop_path;
+  private List<Integer> genre_ids;
+  private String id;
+  private String original_language;
+  private String original_title;
+  private String overview;
+  private String popularity;
+  private String poster_path;
+  private String release_date;
+  private String title;
+  private boolean video;
+  private double vote_average;
+  private int vote_count;
 }
